+++ kaltura-common/src/index.ts
@@ -1,8 +1,8 @@
 export * from './kaltura-common.module';
 export * from './app-storage.service';
-export * from './localization/index';
 export * from './utils/kaltura-utils';
 export * from './upload-management';
 export * from './xml-parser';
 export * from './operation-tag';
 export * from './server-polls'
+export { KalturaLoggerInjectionToken } from './kaltura-logger';