+++ kaltura-ui/src/draggable-data-table/column.component.ts
@@ -1,0 +1,12 @@
+import {Component, ContentChild, Input, TemplateRef} from '@angular/core';
+
+@Component({
+  selector: 'k-column',
+  template: ''
+})
+export class ColumnComponent {
+  @Input() field: string;
+  @Input() style: string;
+  @Input() header: string;
+  @ContentChild(TemplateRef) template: TemplateRef<any>;
+}