+++ kaltura-ui/src/draggable-data-table/index.ts
@@ -1,0 +1,1 @@
+export { DraggableDataTableModule } from "./draggable-data-table.module";