+++ kaltura-logger/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.2.1"></a>
+## [2.2.1](https://github.com/kaltura/kaltura-ng/compare/@kaltura-ng/kaltura-logger@2.2.0...@kaltura-ng/kaltura-logger@2.2.1) (2018-05-31)
+
+### Features
+* import kaltura logger module requires providing root namespace value
+
+### BREAKING CHANGES
+
+before in `app.module.ts`:
+```
+@NgModule({
+  imports: <any>[
+      KalturaLoggerModule.forRoot(),
+  ]
+})
+```
+
+after in `app.module.ts`:
+```
+@NgModule({
+  imports: <any>[
+      KalturaLoggerModule.forRoot('kmc'),
+  ]
+})
+```
+
+
 <a name="2.2.0"></a>
 # [2.2.0](https://github.com/kaltura/kaltura-ng/compare/@kaltura-ng/kaltura-logger@2.1.0...@kaltura-ng/kaltura-logger@2.2.0) (2018-04-30)